fix: grid responsiveness
